fix: treat self-closing "dcl-ds ... end-ds;" as a single-line declaration
A dcl-ds closed by an inline end-ds on the same line (e.g. an EXTNAME
structure) was seen by the block matcher's backward scan as leaving a data
structure open, so a following `if` at the start of a line was misread as a
subfield name and dropped - leaving its `endif` reported as unmatched.

Extract isInsideOpenDclDsBlock into blockParser (making it unit-testable),
and skip a dcl-ds line that also contains an inline end-ds, alongside the
existing likeds()/likerec() case. Add regression tests for the extracted
function.
